+++ ZF/ZFCore/zfsrc/ZFCore/ZFObjectDef/ZFCallbackSerializable.h
@@ -53,6 +53,7 @@ ZF_NAMESPACE_GLOBAL_BEGIN
  * @endcode
  */
 ZFTYPEID_DECLARE(ZFCallback, ZFCallback)
+ZFOUTPUT_TYPE(ZFCallback, {output.execute(v.objectInfo());})
 
 /** @brief keyword for serialize */
 #define ZFSerializableKeyword_ZFCallback_method zfText("method")